ingestworker/pyproject.toml
[project]
name = "ingestworker"
version = "0.1.0"
description = "Add your description here"
readme = "README.md"
requires-python = ">=3.12"
dependencies = [
    "scikit-learn>=1.7.2",
    "timm>=1.0.22",
    "torch>=2.9.1",
    "torchvision>=0.24.1",
    "shared",
    "pillow-heif>=1.1.1",
    "imagehash>=4.3.2",
    "dependency-injector>=4.48.2",
    "celery[redis]>=5.5.3",
]

[tool.uv]
package = false

[tool.uv.sources]
shared = { workspace = true }
